Move the test contract into the unit test class.
